+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated `PersistedPaginatedTable` types
+
 ## [0.6.0] - 2019-12-02
 
 ### Added